Wrap Integration::Session with `while_untenanted`
An integration test must be able to access the database to retrieve
fixtures, or create objects upon which to operate.

But once an HTTP verb is invoked, we want the middleware and the
application code to behave like they would in a live production
app (that is, starting from an untenanted state).