Tenant.connection_pool returns a fake pool when untenanted
Rather than immediately raising NoTenantError, let's return an
UntenantedConnectionPool which we can start to use for
non-connection-related features like schema cache loading.

UntenantedConnectionPool still raises NoTenantError when a connection
is used.

Note that this commit doesn't pass all the tests, because a schema
cache is not being returned. Will be fixed in the next commit.