feat: Allow creating tenants withing transactional tests
Note that the default test tenant, which Rails creates fixtures for,
will still be transactional.

Any other tenants created during a test will not be wrapped in a
transaction, though those other tenant databases will be deleted at
the start of the test suite so the app is always tested from a known
state.