Calling `.tenanted` sets the default shard on the conn class
We want the default shard to be in a state where `current_tenant` will
be nil by default.

Previously we were calling `connecting_to` at include time to put the
connected_to_stack in a state where UNTENANTED_SENTINEL was on the
stack, however that array is thread-local so it wasn't working
correctly in threaded environments (like Active Job workers).

Setting the default_shard class attribute does this in a way that will
work for all threads.