fix: initializers to extend testing frameworks only run in test env
If `app` is called in the console, then the Rails console helper will
load ActionDispatch::Integration::Session to provide URL helpers. That
class loads ActiveSupport::TestCase which will trigger our
active_support_test_case load hook, which will set a current_tenant to
prepare for fixture loading.

Because we only want to set that tenant to prepare for fixture
loading, we should only set up that hook in a test environment. But
more broadly, we really don't need any of our test-framework load
hooks to be invoked unless we're actually testing.