Tenanted databases are no longer implicitly migrated
Previously, when a connection pool was created, we automatically
migrated the database if necessary. We no longer do this, and instead
raise a PendingMigrationError if there are pending migrations.

Tenant.create_tenant is one exception to this behavior: that method
will ensure the new database has an up-to-date schema via schema dump
loading and migration.

Tenant.destroy_tenant is another exception: because we don't care
about the state of the database, that method skips the schema version
check.

ref: https://37s.fizzy.37signals.com/collections/693169862/cards/999008694