General support for tenanted active storage blobs
This approach prefixes the tenant to the value returned by `Blob#key`.

It also simplifies the ActiveStorage::Service extension by dropping
the service-specific `tenanted:` configuration. Note that
ActiveStorage::Record is either tenanted or not, and there's no point
in configuring a service to behave differently from that setting.

Also note a breaking change in DiskService, which is that the path on
disk changes. Developers can choose to have a tenanted root path, but
even if they don't, the blob is still written to a tenant-specific
folder.

ref: https://37s.fizzy.37signals.com/collections/7/cards/246