Ensure tenant context safety on associations.
Because we can't determine ahead of time whether a polymorphic
association will return tenanted models, we assume it will and perform
the context safety check.

It's possible that a polymorphic association will be pointed at
untenanted models, in which case it might be nice to be able to
somehow declare that on the association, maybe like this:

    belongs_to :announceable, polymorphic: true, tenanted: false

but I'm going to wait until someone actually has this use case.

Closes #152