Tenanted::DatabaseTasks is now a class
that must be instantiated with a tenanted database config. This
encapsulates things a bit better as the project heads towards support
for multiple tenanted databases.

The primary outcome of this commit is that `rails db:migrate` will now
run migrations for all tenanted database, where previously it would
just run on the first tenanted database in the configuration.